requirements.txt
 flask==3.0.2
 flasgger>=0.9.5
 flask-cors==4.0.0
 sentence-transformers==2.7.0
 chromadb>=0.4.24
 pyodbc==5.1.0
 numpy==1.26.4
 google-generativeai>=0.3.0